refactor(doclang): isolate source mapping recorder
