+++ docling_core/transforms/deserializer/__init__.py
@@ -1,8 +1,8 @@
 """Define the deserializer types."""
 
 from docling_core.transforms.deserializer.base import BaseDocDeserializer
-from docling_core.transforms.deserializer.doclang import (
-    DocLangDocDeserializer,
+from docling_core.transforms.deserializer.doclang import DocLangDocDeserializer
+from docling_core.transforms.deserializer.doclang_source_mapping import (
     DocLangSourceMap,
     DocLangSourceTarget,
 )
