+++ docling_core/transforms/deserializer/doclang.py
@@ -4,7 +4,6 @@ Aligned to the DocLang specification version ``_DOCLANG_VERSION``.
 """
 
 from collections.abc import Callable, Sequence
-from dataclasses import dataclass, field
 from itertools import groupby
 from pathlib import Path
 from typing import Any, ClassVar, Literal, Optional, cast
@@ -17,6 +16,11 @@ from pydantic import AnyUrl, BaseModel, PrivateAttr
 from typing_extensions import override
 
 from docling_core.transforms.deserializer.base import BaseDocDeserializer
+from docling_core.transforms.deserializer.doclang_source_mapping import (
+    DocLangSourceMap,
+    DocLangSourceTarget,
+    _DocLangSourceRecorder,
+)
 from docling_core.transforms.serializer._doclang_utils import (
     _DOCLANG_META_TAG_DESCRIPTION,
     _DOCLANG_META_TAG_SMILES,
@@ -94,35 +98,6 @@ from docling_core.utils.settings import settings
 __all__ = ["DocLangDocDeserializer", "DocLangSourceMap", "DocLangSourceTarget"]
 
 
-@dataclass(frozen=True)
-class DocLangSourceTarget:
-    """Semantic target created from one DocLang source element."""
-
-    kind: Literal["item", "table_cell", "page"]
-    item_ref: Optional[str] = None
-    row: Optional[int] = None
-    col: Optional[int] = None
-    page_no: Optional[int] = None
-
-
-@dataclass
-class DocLangSourceMap:
-    """Bidirectional XPath-to-semantic bindings collected during deserialization."""
-
-    targets_by_xpath: dict[str, DocLangSourceTarget] = field(default_factory=dict)
-    xpaths_by_target: dict[DocLangSourceTarget, list[str]] = field(default_factory=dict)
-
-    def bind(self, xpath: str, target: DocLangSourceTarget) -> None:
-        """Bind a canonical DocLang XPath to a semantic target."""
-        previous = self.targets_by_xpath.get(xpath)
-        if previous == target:
-            return
-        if previous is not None:
-            self.xpaths_by_target[previous].remove(xpath)
-        self.targets_by_xpath[xpath] = target
-        self.xpaths_by_target.setdefault(target, []).append(xpath)
-
-
 def _utf8_byte_length(text: str) -> int:
     """Return UTF-8 byte length of ``text`` without retaining the encoded buffer."""
     # encode builds a temporary bytes object; acceptable for a one-shot gate check.
@@ -167,8 +142,8 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
     _max_xml_bytes: int = PrivateAttr(default=settings.max_doclang_xml_bytes)
     _max_xml_depth: int = PrivateAttr(default=settings.max_doclang_xml_depth)
     _max_xml_elements: int = PrivateAttr(default=settings.max_doclang_xml_elements)
-    _source_map: Optional[DocLangSourceMap] = PrivateAttr(default=None)
-    _source_binding_suspended: bool = PrivateAttr(default=False)
+    # Optional sidecar: deserialization output must not depend on recorded bindings.
+    _source_recorder: _DocLangSourceRecorder = PrivateAttr(default_factory=_DocLangSourceRecorder)
 
     def _thread_registry_key(self, *, thread_id: str, host: str) -> tuple[str, str]:
         return (thread_id, host)
@@ -208,7 +183,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
             max_xml_bytes: Optional override for ``settings.max_doclang_xml_bytes``.
             max_xml_depth: Optional override for ``settings.max_doclang_xml_depth``.
             max_xml_elements: Optional override for ``settings.max_doclang_xml_elements``.
-            source_map: Optional collector populated with XPath-to-semantic bindings.
+            source_map: Optional recorder output populated with XPath-to-semantic bindings.
 
         Returns:
             A populated `DoclingDocument` parsed from the input.
@@ -219,8 +194,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
         self._max_xml_bytes = int(kwargs.get("max_xml_bytes", settings.max_doclang_xml_bytes))
         self._max_xml_depth = int(kwargs.get("max_xml_depth", settings.max_doclang_xml_depth))
         self._max_xml_elements = int(kwargs.get("max_xml_elements", settings.max_doclang_xml_elements))
-        self._source_map = kwargs.get("source_map")
-        self._source_binding_suspended = False
+        self._source_recorder = _DocLangSourceRecorder(kwargs.get("source_map"))
 
         root = self._parse_xml_string(text)
         if root.tagName != DocLangToken.DOCUMENT.value:
@@ -235,79 +209,6 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
         self._ensure_page_exists(doc=doc, page_no=self._page_no, resolution=self._default_resolution)
         self._parse_document_root(doc=doc, root=root)
         return doc
-
-    @staticmethod
-    def _source_xpath(el: Element) -> str:
-        """Return a canonical namespace-independent XPath using the public ``d`` prefix."""
-        parts: list[str] = []
-        current: Optional[Node] = el
-        while isinstance(current, Element):
-            name = current.localName or current.tagName.rsplit(":", maxsplit=1)[-1]
-            parent = current.parentNode
-            if isinstance(parent, Element):
-                siblings = [
-                    child
-                    for child in parent.childNodes
-                    if isinstance(child, Element)
-                    and (child.localName or child.tagName.rsplit(":", maxsplit=1)[-1]) == name
-                ]
-                parts.append(f"d:{name}[{siblings.index(current) + 1}]")
-            else:
-                parts.append(f"d:{name}")
-            current = parent
-        return "/" + "/".join(reversed(parts))
-
-    def _bind_item(self, el: Element, item: NodeItem) -> None:
-        if self._source_map is not None and not self._source_binding_suspended:
-            self._source_map.bind(
-                self._source_xpath(el),
-                DocLangSourceTarget(kind="item", item_ref=item.self_ref),
-            )
-
-    def _bind_page(self, el: Element) -> None:
-        if self._source_map is not None and not self._source_binding_suspended:
-            self._source_map.bind(
-                self._source_xpath(el),
-                DocLangSourceTarget(kind="page", page_no=self._page_no),
-            )
-
-    def _bind_table_cells(
-        self,
-        el: Element,
-        data: TableData,
-        item_ref: str,
-        *,
-        row_offset: int = 0,
-        col_offset: int = 0,
-    ) -> None:
-        if self._source_map is None or self._source_binding_suspended:
-            return
-        marker_tags = self._OTSL_STRUCTURAL_TAGS - {DocLangToken.NL.value}
-        row = 0
-        col = 0
-        xpath_by_position: dict[tuple[int, int], str] = {}
-        _, body_nodes = self._split_element_children_head_body(el)
-        for node in body_nodes:
-            if not isinstance(node, Element):
-                continue
-            if node.tagName == DocLangToken.NL.value:
-                row += 1
-                col = 0
-            elif node.tagName in marker_tags:
-                xpath_by_position[(row + row_offset, col + col_offset)] = self._source_xpath(node)
-                col += 1
-        for cell in data.table_cells:
-            position = (cell.start_row_offset_idx, cell.start_col_offset_idx)
-            if xpath := xpath_by_position.get(position):
-                self._source_map.bind(
-                    xpath,
-                    DocLangSourceTarget(
-                        kind="table_cell",
-                        item_ref=item_ref,
-                        row=position[0],
-                        col=position[1],
-                    ),
-                )
 
     def _extract_thread_id_from_nodes(self, nodes: Sequence[Node]) -> Optional[str]:
         """Read ``thread_id`` from a ``<thread/>`` element in a node sequence."""
@@ -347,7 +248,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
                 provs.extend(self._provenance_from_location_nodes(doc=doc, nodes=batch))
                 batch = []
                 self._advance_page_break(doc=doc)
-                self._bind_page(node)
+                self._source_recorder.bind_page(node, self._page_no)
             elif isinstance(node, Element) and node.tagName == DocLangToken.LOCATION.value:
                 batch.append(node)
         provs.extend(self._provenance_from_location_nodes(doc=doc, nodes=batch))
@@ -422,7 +323,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
             # Start a new page; keep a default square page using the configured resolution
             self._page_no += 1
             self._ensure_page_exists(doc=doc, page_no=self._page_no, resolution=self._default_resolution)
-            self._bind_page(el)
+            self._source_recorder.bind_page(el, self._page_no)
         elif name == DocLangToken.HEADING.value:
             self._parse_heading(doc=doc, el=el, parent=parent)
         elif name == DocLangToken.FIELD_HEADING.value:
@@ -532,7 +433,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
             ):
                 if prov_list:
                     self._merge_threaded_text_item(text="", prov_list=prov_list, existing=existing)
-                self._bind_item(el, existing)
+                self._source_recorder.bind_item(el, existing)
             return
 
         nm = el.tagName
@@ -548,7 +449,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
                 and isinstance(existing, CodeItem)
             ):
                 self._merge_threaded_text_item(text=code_text, prov_list=prov_list, existing=existing)
-                self._bind_item(el, existing)
+                self._source_recorder.bind_item(el, existing)
                 return
             item = doc.add_code(
                 text=code_text,
@@ -623,7 +524,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
                 and isinstance(existing, TextItem)
             ):
                 self._merge_threaded_text_item(text=text, prov_list=prov_list, existing=existing)
-                self._bind_item(el, existing)
+                self._source_recorder.bind_item(el, existing)
                 return
             item = doc.add_text(
                 label=label,
@@ -645,7 +546,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
                 and isinstance(existing, FormulaItem)
             ):
                 self._merge_threaded_text_item(text=text, prov_list=prov_list, existing=existing)
-                self._bind_item(el, existing)
+                self._source_recorder.bind_item(el, existing)
                 return
             item = doc.add_formula(
                 text=text,
@@ -660,7 +561,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
         else:
             return
 
-        self._bind_item(el, item)
+        self._source_recorder.bind_item(el, item)
 
     def _extract_code_content_and_language(self, el: Element) -> tuple[str, CodeLanguageLabel]:
         """Extract code content and language from a <code> element."""
@@ -710,7 +611,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
                 and isinstance(existing, TextItem)
             ):
                 self._merge_threaded_text_item(text=text_stripped, prov_list=prov_list, existing=existing)
-                self._bind_item(el, existing)
+                self._source_recorder.bind_item(el, existing)
                 return
             # Level 1 maps to TitleItem, level > 1 maps to SectionHeaderItem with level-1
             if level == 1:
@@ -732,7 +633,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
             if thread_id:
                 self._register_thread(thread_id=thread_id, host=DocLangToken.HEADING.value, item=item)
             self._apply_custom_meta_from_element(item=item, el=el)
-            self._bind_item(el, item)
+            self._source_recorder.bind_item(el, item)
 
     def _parse_field_heading(self, *, doc: DoclingDocument, el: Element, parent: Optional[NodeItem]) -> None:
         lvl_txt = el.getAttribute(DocLangAttributeKey.LEVEL.value) or "1"
@@ -752,7 +653,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
                 and isinstance(existing, TextItem)
             ):
                 self._merge_threaded_text_item(text=text_stripped, prov_list=prov_list, existing=existing)
-                self._bind_item(el, existing)
+                self._source_recorder.bind_item(el, existing)
                 return
             item = doc.add_field_heading(
                 text=text_stripped,
@@ -764,7 +665,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
             self._apply_initial_text_provenance(item, text=text_stripped, prov_list=prov_list)
             if thread_id:
                 self._register_thread(thread_id=thread_id, host=DocLangToken.FIELD_HEADING.value, item=item)
-            self._bind_item(el, item)
+            self._source_recorder.bind_item(el, item)
 
     _FIELD_INLINE_BODY_TAGS: ClassVar[frozenset[str]] = frozenset(
         {
@@ -864,7 +765,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
         )
         for prov in prov_list[1:]:
             fri.prov.append(prov)
-        self._bind_item(el, fri)
+        self._source_recorder.bind_item(el, fri)
         _, body_nodes = self._split_element_children_head_body(el)
         self._dispatch_body_nodes(doc=doc, body_nodes=body_nodes, parent=fri)
 
@@ -878,7 +779,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
         )
         for prov in prov_list[1:]:
             fi.prov.append(prov)
-        self._bind_item(el, fi)
+        self._source_recorder.bind_item(el, fi)
         _, body_nodes = self._split_element_children_head_body(el)
         self._dispatch_body_nodes(doc=doc, body_nodes=body_nodes, parent=fi)
 
@@ -895,7 +796,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
         else:
             label = DocItemLabel.CHECKBOX_UNSELECTED
         item = doc.add_text(label=label, text="", parent=parent)
-        self._bind_item(el, item)
+        self._source_recorder.bind_item(el, item)
 
     def _parse_field_hint(self, *, doc: DoclingDocument, el: Element, parent: Optional[NodeItem]) -> None:
         prov_list = self._extract_provenance(doc=doc, el=el)
@@ -912,7 +813,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
             formatting=formatting,
         )
         self._apply_initial_text_provenance(item, text=text_stripped, prov_list=prov_list)
-        self._bind_item(el, item)
+        self._source_recorder.bind_item(el, item)
 
     def _field_kv_needs_inline_container(self, body_nodes: Sequence[Node]) -> bool:
         """True when key/value body must become an inline group, not flat text."""
@@ -965,7 +866,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
                 )
             self._apply_initial_text_provenance(item, text=text, prov_list=prov_list)
             self._apply_custom_meta_from_element(item=item, el=el)
-            self._bind_item(el, item)
+            self._source_recorder.bind_item(el, item)
             return
 
         if needs_inline:
@@ -992,7 +893,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
             )
             self._apply_initial_text_provenance(item, text="", prov_list=prov_list)
             self._apply_custom_meta_from_element(item=item, el=el)
-            self._bind_item(el, item)
+            self._source_recorder.bind_item(el, item)
             return
 
         if is_value:
@@ -1017,7 +918,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
                 doc.add_text(label=DocItemLabel.TEXT, text=node.data.strip(), parent=item)
         self._apply_initial_text_provenance(item, text="", prov_list=prov_list)
         self._apply_custom_meta_from_element(item=item, el=el)
-        self._bind_item(el, item)
+        self._source_recorder.bind_item(el, item)
 
     def _first_non_whitespace_node(self, nodes: Sequence[Node]) -> Optional[Node]:
         """Return the first node that is not whitespace-only text."""
@@ -1317,7 +1218,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
             li_group = doc.add_list_group(parent=parent)
             if thread_id:
                 self._register_thread(thread_id=thread_id, host=DocLangToken.LIST.value, item=li_group)
-        self._bind_item(el, li_group)
+        self._source_recorder.bind_item(el, li_group)
         actual_children = [
             ch for ch in el.childNodes if isinstance(ch, Element) and ch.tagName not in {DocLangToken.LOCATION.value}
         ]
@@ -1339,6 +1240,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
         for start, end in ranges:
             # The ldiv element itself
             ldiv_el = actual_children[start]
+            # Source-map-only: identify the ListItem created by this delimiter.
             child_count = len(li_group.children)
 
             # Extract marker if present within the ldiv
@@ -1491,7 +1393,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
             if len(li_group.children) > child_count:
                 item = li_group.children[child_count].resolve(doc)
                 if isinstance(item, ListItem):
-                    self._bind_item(ldiv_el, item)
+                    self._source_recorder.bind_item(ldiv_el, item)
 
     # ------------- Inline groups -------------
     def _parse_inline_group(
@@ -1506,7 +1408,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
         # Create the inline group
         inline_group = doc.add_inline_group(parent=parent)
         if nodes is None:
-            self._bind_item(el, inline_group)
+            self._source_recorder.bind_item(el, inline_group)
 
         # Process all child elements, adding them as children of the inline group
         my_nodes = nodes or el.childNodes
@@ -1568,7 +1470,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
             and (existing := self._get_thread_item(thread_id, host=table_host)) is not None
             and isinstance(existing, TableItem)
         ):
-            self._bind_item(otsl_el, existing)
+            self._source_recorder.bind_item(otsl_el, existing)
             row_offset, col_offset = _thread_table_merge_offset(existing, tbl_provs[0]) if tbl_provs else (0, 0)
             for prov in tbl_provs:
                 existing.prov.append(prov)
@@ -1582,8 +1484,9 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
                     row_offset=row_offset,
                     col_offset=col_offset,
                 )
-                self._bind_table_cells(
+                self._source_recorder.bind_table_cells(
                     otsl_el,
+                    body_nodes,
                     fragment_td,
                     existing.self_ref,
                     row_offset=row_offset,
@@ -1608,11 +1511,11 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
             content_layer=content_layer,
             label=table_label,
         )
-        self._bind_item(otsl_el, tbl)
+        self._source_recorder.bind_item(otsl_el, tbl)
         tbl_content = _wrap(text=inner, wrap_tag=DocLangToken.TABLE.value)
         td = self._parse_otsl_table_content(otsl_content=tbl_content, doc=doc, parent=tbl)
         tbl.data = td
-        self._bind_table_cells(otsl_el, td, tbl.self_ref)
+        self._source_recorder.bind_table_cells(otsl_el, body_nodes, td, tbl.self_ref)
         for p in tbl_provs[1:]:
             tbl.prov.append(p)
         if thread_id:
@@ -1647,7 +1550,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
             content_layer=content_layer,
         )
         if picture_el is not None:
-            self._bind_item(picture_el, pic)
+            self._source_recorder.bind_item(picture_el, pic)
         for p in prov_list[1:]:
             pic.prov.append(p)
         for ftn in footnotes:
@@ -1694,7 +1597,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
                 _, otsl_body_nodes = self._split_element_children_head_body(node)
                 inner = self._nodes_to_xml(otsl_body_nodes)
                 td = self._parse_otsl_table_content(_wrap(inner, DocLangToken.TABULAR.value))
-                self._bind_table_cells(node, td, pic.self_ref)
+                self._source_recorder.bind_table_cells(node, otsl_body_nodes, td, pic.self_ref)
                 if pic.meta is None:
                     pic.meta = PictureMeta()
                 pic.meta.tabular_chart = TabularChartMetaField(chart_data=td)
@@ -1834,7 +1737,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
         )
         for p in prov_list[1:]:
             item.prov.append(p)
-        self._bind_item(cap_el, item)
+        self._source_recorder.bind_item(cap_el, item)
         return item
 
     def _extract_footnotes(self, *, doc: DoclingDocument, el: Element) -> list[TextItem]:
@@ -1851,7 +1754,7 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
                     )
                     for p in prov_list[1:]:
                         item.prov.append(p)
-                    self._bind_item(node, item)
+                    self._source_recorder.bind_item(node, item)
                     footnotes.append(item)
         return footnotes
 
@@ -2117,12 +2020,8 @@ class DocLangDocDeserializer(BaseDocDeserializer, BaseModel):
                         cell_group = doc.add_group(parent=parent, label=GroupLabel.UNSPECIFIED)
                         text_parts: list[str] = []
                         for part in element_parts:
-                            was_suspended = self._source_binding_suspended
-                            self._source_binding_suspended = True
-                            try:
+                            with self._source_recorder.suspend():
                                 self._dispatch_element(doc=doc, el=part, parent=cell_group)
-                            finally:
-                                self._source_binding_suspended = was_suspended
                             text_parts.append(self._get_text(part))
                         actual_text = "".join(text_parts).strip() or cell_text_stripped
                         table_cells.append(
