+++ test/test_deserializer_doclang.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import pytest
 
-from docling_core.transforms.deserializer.doclang import DocLangDocDeserializer, DocLangSourceMap
+from docling_core.transforms.deserializer.doclang import DocLangDocDeserializer
 from docling_core.transforms.serializer.doclang import (
     DocLangDocSerializer,
     DocLangParams,
@@ -2286,25 +2286,3 @@ def test_deserialize_accepts_document_within_budgets() -> None:
     )
     assert len(doc.texts) == 1
     assert doc.texts[0].text == "hello"
-
-
-def test_deserialize_collects_source_bindings() -> None:
-    xml = """<doclang>
-  <heading level="2">Section</heading>
-  <text><thread thread_id="7"/>first</text>
-  <text><thread thread_id="7"/>second</text>
-  <list><ldiv/><content>item</content></list>
-  <table><ched/>Head<fcel/>Value<nl/></table>
-</doclang>"""
-    source_map = DocLangSourceMap()
-    doc = DocLangDocDeserializer().deserialize_str(xml, source_map=source_map)
-
-    first_thread = source_map.targets_by_xpath["/d:doclang/d:text[1]"]
-    assert first_thread == source_map.targets_by_xpath["/d:doclang/d:text[2]"]
-    assert source_map.targets_by_xpath["/d:doclang/d:list[1]/d:ldiv[1]"].kind == "item"
-    assert source_map.targets_by_xpath["/d:doclang/d:table[1]/d:ched[1]"].row == 0
-    assert source_map.targets_by_xpath["/d:doclang/d:table[1]/d:fcel[1]"].col == 1
-
-    bindings = dict(source_map.targets_by_xpath)
-    doc.hierarchize()
-    assert source_map.targets_by_xpath == bindings
